third_party/WebKit/LayoutTests/http/tests/websocket/workers/shared-worker-cookie-ws-to-ws.html

<!DOCTYPE html>
<script src="/js-test-resources/testharness.js"></script>
<script src="/js-test-resources/testharnessreport.js"></script>
<script>

var test = async_test("Test that cookies set by one WebSocket on opening handshake are sent by another WebSocket on opening handshake");
test.step(function()
{
    var worker = new SharedWorker("resources/cookie-ws-to-ws.js", "websocketCookie");
    worker.port.onmessage = test.step_func(function (evt)
    {
        assert_equals(evt.data, "DONE", "evt.data");
        test.done();
    });
    worker.onerror = test.unreached_func("onerror event should not fire");
});

</script>
